+++ packages/worker-threads-pool/src/index.ts
@@ -1,3 +1,4 @@
 export { registerWorker } from "./registerWorker";
 export { WorkerPool } from "./WorkerPool";
-export type { Pool } from "./Pool";
+export { AggregatedPool } from "./AggregatedPool";
+export type { Pool } from "./types/Pool";