+++ packages/worker-threads-pool/src/types/WorkerPoolOptions.ts
@@ -1,0 +1,7 @@
+import type { WorkerOptions } from "worker_threads";
+
+export interface WorkerPoolOptions {
+  maxWorkers?: number;
+  script: string;
+  workerOptions?: WorkerOptions;
+}